log-server: refactor store types
This in preparation for future new stores.
