+++ packages/log-server/__tests__/app-runs.test.ts
@@ -2,7 +2,8 @@
 
 import { describe, vi, test as vitestTest } from 'vitest';
 import { apiMediaType } from '../src/app-utils.ts';
-import { StoreError, type RunStatus } from '../src/store.js';
+import { StoreError } from '../src/store-errors.ts';
+import type { RunStatus } from '../src/store.ts';
 import { arrayify, firstStrict } from '../src/utils.js';
 import {
   apiContentTypeRegExp,
