+++ packages/log-server/src/app-logs-handlers.ts
@@ -11,7 +11,9 @@ import {
   type SubServerDescription,
 } from './app-utils.js';
 import { csvExportStream } from './csv-export.js';
-import { StoreError, type AllFilter, type Store } from './store.js';
+import { StoreError } from './store-errors.ts';
+import type { AllFilter } from './store-filters.ts';
+import type { DataStore } from './store.ts';
 import { arrayify, firstStrict } from './utils.js';
 
 export const logHandlers = (): SubServerDescription<'/logs'> => ({
@@ -183,7 +185,7 @@ function getResponseMimeType(
 type LogResponseMimeType = 'json' | 'csv';
 
 function jsonResponseStream(
-  store: Store,
+  store: DataStore,
   filter: Omit<AllFilter, 'runStatus'> = {},
   includes: { run?: boolean; experiment?: boolean; lastLogs?: boolean } = {},
 ): Readable {
@@ -191,7 +193,7 @@ function jsonResponseStream(
 }
 
 async function* jsonResponseChunkGenerator(
-  store: Store,
+  store: DataStore,
   filter: Omit<AllFilter, 'runStatus'>,
   includes: { run?: boolean; experiment?: boolean; lastLogs?: boolean },
 ) {
