+++ packages/log-server/src/csv-export.ts
@@ -1,7 +1,8 @@
 import { stringify } from 'csv';
 import { pipeline, Readable } from 'node:stream';
 import { mapKeys, pickBy, pipe } from 'remeda';
-import { type AllFilter, type Log, SQLiteStore as Store } from './store.js';
+import type { AllFilter } from './store-filters.ts';
+import type { DataStore, Log } from './store.ts';
 import { withSnakeCaseProps } from './utils.js';
 
 const csvLogColumns: Array<keyof Log> = [
@@ -13,7 +14,7 @@ const csvLogColumns: Array<keyof Log> = [
 const renamedLogColumns: Partial<Record<keyof Log, string>> = {};
 
 export function csvExportStream(
-  store: Pick<Store, 'getLogs' | 'getLogValueNames'>,
+  store: DataStore,
   filter: Omit<AllFilter, 'runStatus'> = {},
 ): Readable {
   const filterWithValidRun = { ...filter, runStatus: '-canceled' } as const;
