+++ packages/log-server/src/index.ts
@@ -1,2 +1,2 @@
 export { LogServer } from './app.js';
-export { SQLiteStore } from './store.js';
+export { SQLiteStore } from './sqlite-store.ts';
